Crate google_logging2

This documentation was generated from Logging crate version 2.0.0+20210325, where 20210325 is the exact revision of the logging:v2 schema built by the mako code generator v2.0.0.

Everything else about the Logging v2 API can be found at the official documentation site. The original source code is on github.

Features

Handle the following Resources with ease from the central hub ...

billing accounts

buckets get, buckets views get, exclusions create, exclusions delete, exclusions get, exclusions list, exclusions patch, locations buckets create, locations buckets delete, locations buckets list, locations buckets patch, locations buckets undelete, locations buckets views create, locations buckets views delete, locations buckets views list, locations buckets views patch, locations get, locations list, logs delete, logs list, sinks create, sinks delete, sinks get, sinks list, sinks patch and sinks update

entries

list, tail and write

exclusions

create, delete, get, list and patch

folders

exclusions create, exclusions delete, exclusions get, exclusions list, exclusions patch, locations buckets create, locations buckets delete, locations buckets get, locations buckets list, locations buckets patch, locations buckets undelete, locations buckets views create, locations buckets views delete, locations buckets views get, locations buckets views list, locations buckets views patch, locations get, locations list, logs delete, logs list, sinks create, sinks delete, sinks get, sinks list, sinks patch and sinks update

locations

buckets create, buckets delete, buckets get, buckets list, buckets patch, buckets undelete, buckets views create, buckets views delete, buckets views get, buckets views list, buckets views patch, get and list

logs

delete and list

monitored resource descriptors

list

organizations

exclusions create, exclusions delete, exclusions get, exclusions list, exclusions patch, get cmek settings, locations buckets create, locations buckets delete, locations buckets get, locations buckets list, locations buckets patch, locations buckets undelete, locations buckets views create, locations buckets views delete, locations buckets views get, locations buckets views list, locations buckets views patch, locations get, locations list, logs delete, logs list, sinks create, sinks delete, sinks get, sinks list, sinks patch, sinks update and update cmek settings

projects

exclusions create, exclusions delete, exclusions get, exclusions list, exclusions patch, locations buckets create, locations buckets delete, locations buckets get, locations buckets list, locations buckets patch, locations buckets undelete, locations buckets views create, locations buckets views delete, locations buckets views get, locations buckets views list, locations buckets views patch, locations get, locations list, logs delete, logs list, metrics create, metrics delete, metrics get, metrics list, metrics update, sinks create, sinks delete, sinks get, sinks list, sinks patch and sinks update

sinks

create, delete, get, list and update

Other activities are ...

get cmek settings

update cmek settings

Not what you are looking for ? Find all other Google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

creates Method Builders which in turn allow access to individual Call Builders

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

All structures are marked with applicable traits to further categorize them and ease browsing.

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit().await

Or specifically ...

let r = hub.billing_accounts().exclusions_delete(...).doit().await
let r = hub.billing_accounts().locations_buckets_views_delete(...).doit().await
let r = hub.billing_accounts().locations_buckets_delete(...).doit().await
let r = hub.billing_accounts().locations_buckets_undelete(...).doit().await
let r = hub.billing_accounts().logs_delete(...).doit().await
let r = hub.billing_accounts().sinks_delete(...).doit().await
let r = hub.exclusions().delete(...).doit().await
let r = hub.folders().exclusions_delete(...).doit().await
let r = hub.folders().locations_buckets_views_delete(...).doit().await
let r = hub.folders().locations_buckets_delete(...).doit().await
let r = hub.folders().locations_buckets_undelete(...).doit().await
let r = hub.folders().logs_delete(...).doit().await
let r = hub.folders().sinks_delete(...).doit().await
let r = hub.locations().buckets_views_delete(...).doit().await
let r = hub.locations().buckets_delete(...).doit().await
let r = hub.locations().buckets_undelete(...).doit().await
let r = hub.logs().delete(...).doit().await
let r = hub.organizations().exclusions_delete(...).doit().await
let r = hub.organizations().locations_buckets_views_delete(...).doit().await
let r = hub.organizations().locations_buckets_delete(...).doit().await
let r = hub.organizations().locations_buckets_undelete(...).doit().await
let r = hub.organizations().logs_delete(...).doit().await
let r = hub.organizations().sinks_delete(...).doit().await
let r = hub.projects().exclusions_delete(...).doit().await
let r = hub.projects().locations_buckets_views_delete(...).doit().await
let r = hub.projects().locations_buckets_delete(...).doit().await
let r = hub.projects().locations_buckets_undelete(...).doit().await
let r = hub.projects().logs_delete(...).doit().await
let r = hub.projects().metrics_delete(...).doit().await
let r = hub.projects().sinks_delete(...).doit().await
let r = hub.sinks().delete(...).doit().await

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

Setting up your Project

To use this library, you would put the following lines into your Cargo.toml file:

[dependencies]
google-logging2 = "*"
# This project intentionally uses an old version of Hyper. See
# https://github.com/Byron/google-apis-rs/issues/173 for more
# information.
hyper = "^0.14"
hyper-rustls = "^0.22"
serde = "^1.0"
serde_json = "^1.0"
yup-oauth2 = "^5.0"

A complete example

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_logging2 as logging2;
use logging2::api::UndeleteBucketRequest;
use logging2::{Result, Error};
use std::default::Default;
use oauth2;
use logging2::Logging;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: oauth2::ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = yup_oauth2::InstalledFlowAuthenticator::builder(
        secret,
        yup_oauth2::InstalledFlowReturnMethod::HTTPRedirect,
    ).build().await.unwrap();
let mut hub = Logging::new(hyper::Client::builder().build(hyper_rustls::HttpsConnector::with_native_roots()), auth);
// As the method needs a request, you would usually fill it with the desired information
// into the respective structure. Some of the parts shown here might not be applicable !
// Values shown here are possibly random and not representative !
let mut req = UndeleteBucketRequest::default();
 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.billing_accounts().locations_buckets_undelete(req, "name")
             .doit().await;
 
match result {
    Err(e) => match e {
        // The Error enum provides details about what exactly happened.
        // You can also just use its `Debug`, `Display` or `Error` traits
         Error::HttpError(_)
        |Error::Io(_)
        |Error::MissingAPIKey
        |Error::MissingToken(_)
        |Error::Cancelled
        |Error::UploadSizeLimitExceeded(_, _)
        |Error::Failure(_)
        |Error::BadRequest(_)
        |Error::FieldClash(_)
        |Error::JsonDecodeError(_, _) => println!("{}", e),
    },
    Ok(res) => println!("Success: {:?}", res),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

Uploads and Downloads

If a method supports downloads, the response body, which is part of the Result, should be read by you to obtain the media. If such a method also supports a Response Result, it will return that by default. You can see it as meta-data for the actual media. To trigger a media download, you will have to set up the builder by making this call: .param("alt", "media").

Methods supporting uploads can do so using up to 2 different protocols: simple and resumable. The distinctiveness of each is represented by customized doit(...) methods, which are then named upload(...) and upload_resumable(...) respectively.

Customization and Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

Optional Parts in Server-Requests

All structures provided by this library are made to be encodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifiable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are moved

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Re-exports

pub use api::Logging;

pub use client::Result;

pub use client::Error;

pub use client::Delegate;

Modules

api

client