Struct google_logging2::api::FolderMethods

pub struct FolderMethods<'a, C> where
    C: 'a,  { /* fields omitted */ }

A builder providing access to all methods supported on folder resources. It is not used directly, but through the Logging hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_logging2 as logging2;
 
use std::default::Default;
use oauth2;
use logging2::Logging;
 
let secret: oauth2::ApplicationSecret = Default::default();
let auth = yup_oauth2::InstalledFlowAuthenticator::builder(
        secret,
        yup_oauth2::InstalledFlowReturnMethod::HTTPRedirect,
    ).build().await.unwrap();
let mut hub = Logging::new(hyper::Client::builder().build(hyper_rustls::HttpsConnector::with_native_roots()), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `exclusions_create(...)`, `exclusions_delete(...)`, `exclusions_get(...)`, `exclusions_list(...)`, `exclusions_patch(...)`, `locations_buckets_create(...)`, `locations_buckets_delete(...)`, `locations_buckets_get(...)`, `locations_buckets_list(...)`, `locations_buckets_patch(...)`, `locations_buckets_undelete(...)`, `locations_buckets_views_create(...)`, `locations_buckets_views_delete(...)`, `locations_buckets_views_get(...)`, `locations_buckets_views_list(...)`, `locations_buckets_views_patch(...)`, `locations_get(...)`, `locations_list(...)`, `logs_delete(...)`, `logs_list(...)`, `sinks_create(...)`, `sinks_delete(...)`, `sinks_get(...)`, `sinks_list(...)`, `sinks_patch(...)` and `sinks_update(...)`
// to build up your call.
let rb = hub.folders();

Implementations

impl<'a, C> FolderMethods<'a, C>

pub fn exclusions_create(
&self,
request: LogExclusion,
parent: &str
) -> FolderExclusionCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates a new exclusion in a specified parent resource. Only log entries belonging to that resource can be excluded. You can have up to 10 exclusions in a resource.

Arguments

request - No description provided.

parent - Required. The parent resource in which to create the exclusion: "projects/[PROJECT_ID]" "organizations/[ORGANIZATION_ID]" "billingAccounts/[BILLING_ACCOUNT_ID]" "folders/[FOLDER_ID]" Examples: "projects/my-logging-project", "organizations/123456789".

pub fn exclusions_delete(&self, name: &str) -> FolderExclusionDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes an exclusion.

Arguments

name - Required. The resource name of an existing exclusion to delete: "projects/[PROJECT_ID]/exclusions/[EXCLUSION_ID]" "organizations/[ORGANIZATION_ID]/exclusions/[EXCLUSION_ID]" "billingAccounts/[BILLING_ACCOUNT_ID]/exclusions/[EXCLUSION_ID]" "folders/[FOLDER_ID]/exclusions/[EXCLUSION_ID]" Example: "projects/my-project-id/exclusions/my-exclusion-id".

pub fn exclusions_get(&self, name: &str) -> FolderExclusionGetCall<'a, C>

Create a builder to help you perform the following task:

Gets the description of an exclusion.

Arguments

name - Required. The resource name of an existing exclusion: "projects/[PROJECT_ID]/exclusions/[EXCLUSION_ID]" "organizations/[ORGANIZATION_ID]/exclusions/[EXCLUSION_ID]" "billingAccounts/[BILLING_ACCOUNT_ID]/exclusions/[EXCLUSION_ID]" "folders/[FOLDER_ID]/exclusions/[EXCLUSION_ID]" Example: "projects/my-project-id/exclusions/my-exclusion-id".

pub fn exclusions_list(&self, parent: &str) -> FolderExclusionListCall<'a, C>

Create a builder to help you perform the following task:

Lists all the exclusions in a parent resource.

Arguments

parent - Required. The parent resource whose exclusions are to be listed. "projects/[PROJECT_ID]" "organizations/[ORGANIZATION_ID]" "billingAccounts/[BILLING_ACCOUNT_ID]" "folders/[FOLDER_ID]"

pub fn exclusions_patch(
&self,
request: LogExclusion,
name: &str
) -> FolderExclusionPatchCall<'a, C>

Create a builder to help you perform the following task:

Changes one or more properties of an existing exclusion.

Arguments

request - No description provided.

name - Required. The resource name of the exclusion to update: "projects/[PROJECT_ID]/exclusions/[EXCLUSION_ID]" "organizations/[ORGANIZATION_ID]/exclusions/[EXCLUSION_ID]" "billingAccounts/[BILLING_ACCOUNT_ID]/exclusions/[EXCLUSION_ID]" "folders/[FOLDER_ID]/exclusions/[EXCLUSION_ID]" Example: "projects/my-project-id/exclusions/my-exclusion-id".

pub fn locations_buckets_views_create(
&self,
request: LogView,
parent: &str
) -> FolderLocationBucketViewCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates a view over logs in a bucket. A bucket may contain a maximum of 50 views.

Arguments

request - No description provided.

parent - Required. The bucket in which to create the view "projects/[PROJECT_ID]/locations/[LOCATION_ID]/buckets/[BUCKET_ID]" Example: "projects/my-logging-project/locations/my-location/buckets/my-bucket"

pub fn locations_buckets_views_delete(
&self,
name: &str
) -> FolderLocationBucketViewDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes a view from a bucket.

Arguments

name - Required. The full resource name of the view to delete: "projects/[PROJECT_ID]/locations/[LOCATION_ID]/buckets/[BUCKET_ID]/views/[VIEW_ID]" Example: "projects/my-project-id/locations/my-location/buckets/my-bucket-id/views/my-view-id".

pub fn locations_buckets_views_get(
&self,
name: &str
) -> FolderLocationBucketViewGetCall<'a, C>

Create a builder to help you perform the following task:

Gets a view.

Arguments

name - Required. The resource name of the policy: "projects/[PROJECT_ID]/locations/[LOCATION_ID]/buckets/[BUCKET_ID]/views/[VIEW_ID]" Example: "projects/my-project-id/locations/my-location/buckets/my-bucket-id/views/my-view-id".

pub fn locations_buckets_views_list(
&self,
parent: &str
) -> FolderLocationBucketViewListCall<'a, C>

Create a builder to help you perform the following task:

Lists views on a bucket.

Arguments

parent - Required. The bucket whose views are to be listed: "projects/[PROJECT_ID]/locations/[LOCATION_ID]/buckets/[BUCKET_ID]"

pub fn locations_buckets_views_patch(
&self,
request: LogView,
name: &str
) -> FolderLocationBucketViewPatchCall<'a, C>

Create a builder to help you perform the following task:

Updates a view. This method replaces the following fields in the existing view with values from the new view: filter.

Arguments

request - No description provided.

name - Required. The full resource name of the view to update "projects/[PROJECT_ID]/locations/[LOCATION_ID]/buckets/[BUCKET_ID]/views/[VIEW_ID]" Example: "projects/my-project-id/locations/my-location/buckets/my-bucket-id/views/my-view-id".

pub fn locations_buckets_create(
&self,
request: LogBucket,
parent: &str
) -> FolderLocationBucketCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates a bucket that can be used to store log entries. Once a bucket has been created, the region cannot be changed.

Arguments

request - No description provided.

parent - Required. The resource in which to create the bucket: "projects/[PROJECT_ID]/locations/[LOCATION_ID]" Example: "projects/my-logging-project/locations/global"

pub fn locations_buckets_delete(
&self,
name: &str
) -> FolderLocationBucketDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes a bucket. Moves the bucket to the DELETE_REQUESTED state. After 7 days, the bucket will be purged and all logs in the bucket will be permanently deleted.

Arguments

name - Required. The full resource name of the bucket to delete. "projects/[PROJECT_ID]/locations/[LOCATION_ID]/buckets/[BUCKET_ID]" "organizations/[ORGANIZATION_ID]/locations/[LOCATION_ID]/buckets/[BUCKET_ID]" "billingAccounts/[BILLING_ACCOUNT_ID]/locations/[LOCATION_ID]/buckets/[BUCKET_ID]" "folders/[FOLDER_ID]/locations/[LOCATION_ID]/buckets/[BUCKET_ID]" Example: "projects/my-project-id/locations/my-location/buckets/my-bucket-id".

pub fn locations_buckets_get(
&self,
name: &str
) -> FolderLocationBucketGetCall<'a, C>

Create a builder to help you perform the following task:

Gets a bucket.

Arguments

name - Required. The resource name of the bucket: "projects/[PROJECT_ID]/locations/[LOCATION_ID]/buckets/[BUCKET_ID]" "organizations/[ORGANIZATION_ID]/locations/[LOCATION_ID]/buckets/[BUCKET_ID]" "billingAccounts/[BILLING_ACCOUNT_ID]/locations/[LOCATION_ID]/buckets/[BUCKET_ID]" "folders/[FOLDER_ID]/locations/[LOCATION_ID]/buckets/[BUCKET_ID]" Example: "projects/my-project-id/locations/my-location/buckets/my-bucket-id".

pub fn locations_buckets_list(
&self,
parent: &str
) -> FolderLocationBucketListCall<'a, C>

Create a builder to help you perform the following task:

Lists buckets.

Arguments

parent - Required. The parent resource whose buckets are to be listed: "projects/[PROJECT_ID]/locations/[LOCATION_ID]" "organizations/[ORGANIZATION_ID]/locations/[LOCATION_ID]" "billingAccounts/[BILLING_ACCOUNT_ID]/locations/[LOCATION_ID]" "folders/[FOLDER_ID]/locations/[LOCATION_ID]" Note: The locations portion of the resource must be specified, but supplying the character - in place of LOCATION_ID will return all buckets.

pub fn locations_buckets_patch(
&self,
request: LogBucket,
name: &str
) -> FolderLocationBucketPatchCall<'a, C>

Create a builder to help you perform the following task:

Updates a bucket. This method replaces the following fields in the existing bucket with values from the new bucket: retention_periodIf the retention period is decreased and the bucket is locked, FAILED_PRECONDITION will be returned.If the bucket has a LifecycleState of DELETE_REQUESTED, FAILED_PRECONDITION will be returned.A buckets region may not be modified after it is created.

Arguments

request - No description provided.

name - Required. The full resource name of the bucket to update. "projects/[PROJECT_ID]/locations/[LOCATION_ID]/buckets/[BUCKET_ID]" "organizations/[ORGANIZATION_ID]/locations/[LOCATION_ID]/buckets/[BUCKET_ID]" "billingAccounts/[BILLING_ACCOUNT_ID]/locations/[LOCATION_ID]/buckets/[BUCKET_ID]" "folders/[FOLDER_ID]/locations/[LOCATION_ID]/buckets/[BUCKET_ID]" Example: "projects/my-project-id/locations/my-location/buckets/my-bucket-id". Also requires permission "resourcemanager.projects.updateLiens" to set the locked property

pub fn locations_buckets_undelete(
&self,
request: UndeleteBucketRequest,
name: &str
) -> FolderLocationBucketUndeleteCall<'a, C>

Create a builder to help you perform the following task:

Undeletes a bucket. A bucket that has been deleted may be undeleted within the grace period of 7 days.

Arguments

request - No description provided.

name - Required. The full resource name of the bucket to undelete. "projects/[PROJECT_ID]/locations/[LOCATION_ID]/buckets/[BUCKET_ID]" "organizations/[ORGANIZATION_ID]/locations/[LOCATION_ID]/buckets/[BUCKET_ID]" "billingAccounts/[BILLING_ACCOUNT_ID]/locations/[LOCATION_ID]/buckets/[BUCKET_ID]" "folders/[FOLDER_ID]/locations/[LOCATION_ID]/buckets/[BUCKET_ID]" Example: "projects/my-project-id/locations/my-location/buckets/my-bucket-id".

pub fn locations_get(&self, name: &str) -> FolderLocationGetCall<'a, C>

Create a builder to help you perform the following task:

Gets information about a location.

Arguments

name - Resource name for the location.

pub fn locations_list(&self, name: &str) -> FolderLocationListCall<'a, C>

Create a builder to help you perform the following task:

Lists information about the supported locations for this service.

Arguments

name - The resource that owns the locations collection, if applicable.

pub fn logs_delete(&self, log_name: &str) -> FolderLogDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes all the log entries in a log for the _Default Log Bucket. The log reappears if it receives new entries. Log entries written shortly before the delete operation might not be deleted. Entries received after the delete operation with a timestamp before the operation will be deleted.

Arguments

logName - Required. The resource name of the log to delete: "projects/[PROJECT_ID]/logs/[LOG_ID]" "organizations/[ORGANIZATION_ID]/logs/[LOG_ID]" "billingAccounts/[BILLING_ACCOUNT_ID]/logs/[LOG_ID]" "folders/[FOLDER_ID]/logs/[LOG_ID]" [LOG_ID] must be URL-encoded. For example, "projects/my-project-id/logs/syslog", "organizations/1234567890/logs/cloudresourcemanager.googleapis.com%2Factivity". For more information about log names, see LogEntry.

pub fn logs_list(&self, parent: &str) -> FolderLogListCall<'a, C>

Create a builder to help you perform the following task:

Lists the logs in projects, organizations, folders, or billing accounts. Only logs that have entries are listed.

Arguments

parent - Required. The resource name that owns the logs: "projects/[PROJECT_ID]" "organizations/[ORGANIZATION_ID]" "billingAccounts/[BILLING_ACCOUNT_ID]" "folders/[FOLDER_ID]"

pub fn sinks_create(
&self,
request: LogSink,
parent: &str
) -> FolderSinkCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates a sink that exports specified log entries to a destination. The export of newly-ingested log entries begins immediately, unless the sink's writer_identity is not permitted to write to the destination. A sink can export log entries only from the resource owning the sink.

Arguments

request - No description provided.

parent - Required. The resource in which to create the sink: "projects/[PROJECT_ID]" "organizations/[ORGANIZATION_ID]" "billingAccounts/[BILLING_ACCOUNT_ID]" "folders/[FOLDER_ID]" Examples: "projects/my-logging-project", "organizations/123456789".

pub fn sinks_delete(&self, sink_name: &str) -> FolderSinkDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes a sink. If the sink has a unique writer_identity, then that service account is also deleted.

Arguments

sinkName - Required. The full resource name of the sink to delete, including the parent resource and the sink identifier: "projects/[PROJECT_ID]/sinks/[SINK_ID]" "organizations/[ORGANIZATION_ID]/sinks/[SINK_ID]" "billingAccounts/[BILLING_ACCOUNT_ID]/sinks/[SINK_ID]" "folders/[FOLDER_ID]/sinks/[SINK_ID]" Example: "projects/my-project-id/sinks/my-sink-id".

pub fn sinks_get(&self, sink_name: &str) -> FolderSinkGetCall<'a, C>

Create a builder to help you perform the following task:

Gets a sink.

Arguments

sinkName - Required. The resource name of the sink: "projects/[PROJECT_ID]/sinks/[SINK_ID]" "organizations/[ORGANIZATION_ID]/sinks/[SINK_ID]" "billingAccounts/[BILLING_ACCOUNT_ID]/sinks/[SINK_ID]" "folders/[FOLDER_ID]/sinks/[SINK_ID]" Example: "projects/my-project-id/sinks/my-sink-id".

pub fn sinks_list(&self, parent: &str) -> FolderSinkListCall<'a, C>

Create a builder to help you perform the following task:

Lists sinks.

Arguments

parent - Required. The parent resource whose sinks are to be listed: "projects/[PROJECT_ID]" "organizations/[ORGANIZATION_ID]" "billingAccounts/[BILLING_ACCOUNT_ID]" "folders/[FOLDER_ID]"

pub fn sinks_patch(
&self,
request: LogSink,
sink_name: &str
) -> FolderSinkPatchCall<'a, C>

Create a builder to help you perform the following task:

Updates a sink. This method replaces the following fields in the existing sink with values from the new sink: destination, and filter.The updated sink might also have a new writer_identity; see the unique_writer_identity field.

Arguments

request - No description provided.

sinkName - Required. The full resource name of the sink to update, including the parent resource and the sink identifier: "projects/[PROJECT_ID]/sinks/[SINK_ID]" "organizations/[ORGANIZATION_ID]/sinks/[SINK_ID]" "billingAccounts/[BILLING_ACCOUNT_ID]/sinks/[SINK_ID]" "folders/[FOLDER_ID]/sinks/[SINK_ID]" Example: "projects/my-project-id/sinks/my-sink-id".

pub fn sinks_update(
&self,
request: LogSink,
sink_name: &str
) -> FolderSinkUpdateCall<'a, C>

Create a builder to help you perform the following task:

Updates a sink. This method replaces the following fields in the existing sink with values from the new sink: destination, and filter.The updated sink might also have a new writer_identity; see the unique_writer_identity field.

Arguments

request - No description provided.

sinkName - Required. The full resource name of the sink to update, including the parent resource and the sink identifier: "projects/[PROJECT_ID]/sinks/[SINK_ID]" "organizations/[ORGANIZATION_ID]/sinks/[SINK_ID]" "billingAccounts/[BILLING_ACCOUNT_ID]/sinks/[SINK_ID]" "folders/[FOLDER_ID]/sinks/[SINK_ID]" Example: "projects/my-project-id/sinks/my-sink-id".