Crate google_logging2

This documentation was generated from Logging crate version 1.0.9+20190629, where 20190629 is the exact revision of the logging:v2 schema built by the mako code generator v1.0.9.

Everything else about the Logging v2 API can be found at the official documentation site. The original source code is on github.

Features

Handle the following Resources with ease from the central hub ...

billing accounts

exclusions create, exclusions delete, exclusions get, exclusions list, exclusions patch, logs delete, logs list, sinks create, sinks delete, sinks get, sinks list, sinks patch and sinks update

entries

list and write

exclusions

create, delete, get, list and patch

folders

exclusions create, exclusions delete, exclusions get, exclusions list, exclusions patch, logs delete, logs list, sinks create, sinks delete, sinks get, sinks list, sinks patch and sinks update

logs

delete and list

monitored resource descriptors

list

organizations

exclusions create, exclusions delete, exclusions get, exclusions list, exclusions patch, logs delete, logs list, sinks create, sinks delete, sinks get, sinks list, sinks patch and sinks update

projects

exclusions create, exclusions delete, exclusions get, exclusions list, exclusions patch, logs delete, logs list, metrics create, metrics delete, metrics get, metrics list, metrics update, sinks create, sinks delete, sinks get, sinks list, sinks patch and sinks update

sinks

create, delete, get, list and update

Not what you are looking for ? Find all other Google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

creates Method Builders which in turn allow access to individual Call Builders

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

All structures are marked with applicable traits to further categorize them and ease browsing.

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit()

Or specifically ...

let r = hub.sinks().create(...).doit()
let r = hub.organizations().sinks_patch(...).doit()
let r = hub.projects().sinks_get(...).doit()
let r = hub.billing_accounts().sinks_get(...).doit()
let r = hub.billing_accounts().sinks_create(...).doit()
let r = hub.billing_accounts().sinks_patch(...).doit()
let r = hub.folders().sinks_create(...).doit()
let r = hub.projects().sinks_update(...).doit()
let r = hub.projects().sinks_patch(...).doit()
let r = hub.organizations().sinks_create(...).doit()
let r = hub.organizations().sinks_update(...).doit()
let r = hub.folders().sinks_patch(...).doit()
let r = hub.organizations().sinks_get(...).doit()
let r = hub.projects().sinks_create(...).doit()
let r = hub.sinks().update(...).doit()
let r = hub.folders().sinks_update(...).doit()
let r = hub.sinks().get(...).doit()
let r = hub.billing_accounts().sinks_update(...).doit()
let r = hub.folders().sinks_get(...).doit()

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

Setting up your Project

To use this library, you would put the following lines into your Cargo.toml file:

[dependencies]
google-logging2 = "*"
# This project intentionally uses an old version of Hyper. See
# https://github.com/Byron/google-apis-rs/issues/173 for more
# information.
hyper = "^0.10"
hyper-rustls = "^0.6"
serde = "^1.0"
serde_json = "^1.0"
yup-oauth2 = "^1.0"

A complete example

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_logging2 as logging2;
use logging2::LogSink;
use logging2::{Result, Error};
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use logging2::Logging;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Logging::new(hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())), auth);
// As the method needs a request, you would usually fill it with the desired information
// into the respective structure. Some of the parts shown here might not be applicable !
// Values shown here are possibly random and not representative !
let mut req = LogSink::default();
 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.organizations().sinks_patch(req, "sinkName")
             .update_mask("et")
             .unique_writer_identity(true)
             .doit();
 
match result {
    Err(e) => match e {
        // The Error enum provides details about what exactly happened.
        // You can also just use its `Debug`, `Display` or `Error` traits
         Error::HttpError(_)
        |Error::MissingAPIKey
        |Error::MissingToken(_)
        |Error::Cancelled
        |Error::UploadSizeLimitExceeded(_, _)
        |Error::Failure(_)
        |Error::BadRequest(_)
        |Error::FieldClash(_)
        |Error::JsonDecodeError(_, _) => println!("{}", e),
    },
    Ok(res) => println!("Success: {:?}", res),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

Uploads and Downloads

If a method supports downloads, the response body, which is part of the Result, should be read by you to obtain the media. If such a method also supports a Response Result, it will return that by default. You can see it as meta-data for the actual media. To trigger a media download, you will have to set up the builder by making this call: .param("alt", "media").

Methods supporting uploads can do so using up to 2 different protocols: simple and resumable. The distinctiveness of each is represented by customized doit(...) methods, which are then named upload(...) and upload_resumable(...) respectively.

Customization and Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

Optional Parts in Server-Requests

All structures provided by this library are made to be enocodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifiable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are moved

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Structs

BillingAccountExclusionCreateCall
Creates a new exclusion in a specified parent resource. Only log entries belonging to that resource can be excluded. You can have up to 10 exclusions in a resource.

BillingAccountExclusionDeleteCall
Deletes an exclusion.

BillingAccountExclusionGetCall
Gets the description of an exclusion.

BillingAccountExclusionListCall
Lists all the exclusions in a parent resource.

BillingAccountExclusionPatchCall
Changes one or more properties of an existing exclusion.

BillingAccountLogDeleteCall
Deletes all the log entries in a log. The log reappears if it receives new entries. Log entries written shortly before the delete operation might not be deleted.

BillingAccountLogListCall
Lists the logs in projects, organizations, folders, or billing accounts. Only logs that have entries are listed.

BillingAccountMethods
A builder providing access to all methods supported on billingAccount resources. It is not used directly, but through the Logging hub.

BillingAccountSinkCreateCall
Creates a sink that exports specified log entries to a destination. The export of newly-ingested log entries begins immediately, unless the sink's writer_identity is not permitted to write to the destination. A sink can export log entries only from the resource owning the sink.

BillingAccountSinkDeleteCall
Deletes a sink. If the sink has a unique writer_identity, then that service account is also deleted.

BillingAccountSinkGetCall
Gets a sink.

BillingAccountSinkListCall
Lists sinks.

BillingAccountSinkPatchCall
Updates a sink. This method replaces the following fields in the existing sink with values from the new sink: destination, and filter.The updated sink might also have a new writer_identity; see the unique_writer_identity field.

BillingAccountSinkUpdateCall
Updates a sink. This method replaces the following fields in the existing sink with values from the new sink: destination, and filter.The updated sink might also have a new writer_identity; see the unique_writer_identity field.

BucketOptions
BucketOptions describes the bucket boundaries used to create a histogram for the distribution. The buckets can be in a linear sequence, an exponential sequence, or each bucket can be specified explicitly. BucketOptions does not include the number of values in each bucket.A bucket has an inclusive lower bound and exclusive upper bound for the values that are counted for that bucket. The upper bound of a bucket must be strictly greater than the lower bound. The sequence of N buckets for a distribution consists of an underflow bucket (number 0), zero or more finite buckets (number 1 through N - 2) and an overflow bucket (number N - 1). The buckets are contiguous: the lower bound of bucket i (i > 0) is the same as the upper bound of bucket i - 1. The buckets span the whole range of finite values: lower bound of the underflow bucket is -infinity and the upper bound of the overflow bucket is +infinity. The finite buckets are so-called because both bounds are finite.

Chunk

ContentRange
Implements the Content-Range header, for serialization only

DefaultDelegate
A delegate with a conservative default implementation, which is used if no other delegate is set.

DummyNetworkStream

Empty
A generic empty message that you can re-use to avoid defining duplicated empty messages in your APIs. A typical example is to use it as the request or the response type of an API method. For instance: service Foo { rpc Bar(google.protobuf.Empty) returns (google.protobuf.Empty); } The JSON representation for Empty is empty JSON object {}.

EntryListCall
Lists log entries. Use this method to retrieve log entries that originated from a project/folder/organization/billing account. For ways to export log entries, see Exporting Logs.

EntryMethods
A builder providing access to all methods supported on entry resources. It is not used directly, but through the Logging hub.

EntryWriteCall
Writes log entries to Logging. This API method is the only way to send log entries to Logging. This method is used, directly or indirectly, by the Logging agent (fluentd) and all logging libraries configured to use Logging. A single request may contain log entries for a maximum of 1000 different resources (projects, organizations, billing accounts or folders)

ErrorResponse
A utility to represent detailed errors we might see in case there are BadRequests. The latter happen if the sent parameters or request structures are unsound

ExclusionCreateCall
Creates a new exclusion in a specified parent resource. Only log entries belonging to that resource can be excluded. You can have up to 10 exclusions in a resource.

ExclusionDeleteCall
Deletes an exclusion.

ExclusionGetCall
Gets the description of an exclusion.

ExclusionListCall
Lists all the exclusions in a parent resource.

ExclusionMethods
A builder providing access to all methods supported on exclusion resources. It is not used directly, but through the Logging hub.

ExclusionPatchCall
Changes one or more properties of an existing exclusion.

Explicit
Specifies a set of buckets with arbitrary widths.There are size(bounds) + 1 (= N) buckets. Bucket i has the following boundaries:Upper bound (0 <= i < N-1): boundsi Lower bound (1 <= i < N); boundsi - 1The bounds field must contain at least one element. If bounds has only one element, then there are no finite buckets, and that single element is the common boundary of the overflow and underflow buckets.

Exponential
Specifies an exponential sequence of buckets that have a width that is proportional to the value of the lower bound. Each bucket represents a constant relative uncertainty on a specific value in the bucket.There are num_finite_buckets + 2 (= N) buckets. Bucket i has the following boundaries:Upper bound (0 <= i < N-1): scale * (growth_factor ^ i). Lower bound (1 <= i < N): scale * (growth_factor ^ (i - 1)).

FolderExclusionCreateCall
Creates a new exclusion in a specified parent resource. Only log entries belonging to that resource can be excluded. You can have up to 10 exclusions in a resource.

FolderExclusionDeleteCall
Deletes an exclusion.

FolderExclusionGetCall
Gets the description of an exclusion.

FolderExclusionListCall
Lists all the exclusions in a parent resource.

FolderExclusionPatchCall
Changes one or more properties of an existing exclusion.

FolderLogDeleteCall
Deletes all the log entries in a log. The log reappears if it receives new entries. Log entries written shortly before the delete operation might not be deleted.

FolderLogListCall
Lists the logs in projects, organizations, folders, or billing accounts. Only logs that have entries are listed.

FolderMethods
A builder providing access to all methods supported on folder resources. It is not used directly, but through the Logging hub.

FolderSinkCreateCall
Creates a sink that exports specified log entries to a destination. The export of newly-ingested log entries begins immediately, unless the sink's writer_identity is not permitted to write to the destination. A sink can export log entries only from the resource owning the sink.

FolderSinkDeleteCall
Deletes a sink. If the sink has a unique writer_identity, then that service account is also deleted.

FolderSinkGetCall
Gets a sink.

FolderSinkListCall
Lists sinks.

FolderSinkPatchCall
Updates a sink. This method replaces the following fields in the existing sink with values from the new sink: destination, and filter.The updated sink might also have a new writer_identity; see the unique_writer_identity field.

FolderSinkUpdateCall
Updates a sink. This method replaces the following fields in the existing sink with values from the new sink: destination, and filter.The updated sink might also have a new writer_identity; see the unique_writer_identity field.

HttpRequest
A common proto for logging HTTP requests. Only contains semantics defined by the HTTP specification. Product-specific logging information MUST be defined in a separate message.

JsonServerError
A utility type which can decode a server response that indicates error

LabelDescriptor
A description of a label.

Linear
Specifies a linear sequence of buckets that all have the same width (except overflow and underflow). Each bucket represents a constant absolute uncertainty on the specific value in the bucket.There are num_finite_buckets + 2 (= N) buckets. Bucket i has the following boundaries:Upper bound (0 <= i < N-1): offset + (width * i). Lower bound (1 <= i < N): offset + (width * (i - 1)).

ListExclusionsResponse
Result returned from ListExclusions.

ListLogEntriesRequest
The parameters to ListLogEntries.

ListLogEntriesResponse
Result returned from ListLogEntries.

ListLogMetricsResponse
Result returned from ListLogMetrics.

ListLogsResponse
Result returned from ListLogs.

ListMonitoredResourceDescriptorsResponse
Result returned from ListMonitoredResourceDescriptors.

ListSinksResponse
Result returned from ListSinks.

LogDeleteCall
Deletes all the log entries in a log. The log reappears if it receives new entries. Log entries written shortly before the delete operation might not be deleted.

LogEntry
An individual entry in a log.

LogEntryOperation
Additional information about a potentially long-running operation with which a log entry is associated.

LogEntrySourceLocation
Additional information about the source code location that produced the log entry.

LogExclusion
Specifies a set of log entries that are not to be stored in Logging. If your project receives a large volume of logs, you might be able to use exclusions to reduce your chargeable logs. Exclusions are processed after log sinks, so you can export log entries before they are excluded. Audit log entries and log entries from Amazon Web Services are never excluded.

LogListCall
Lists the logs in projects, organizations, folders, or billing accounts. Only logs that have entries are listed.

LogMethods
A builder providing access to all methods supported on log resources. It is not used directly, but through the Logging hub.

LogMetric
Describes a logs-based metric. The value of the metric is the number of log entries that match a logs filter in a given time interval.Logs-based metric can also be used to extract values from logs and create a a distribution of the values. The distribution records the statistics of the extracted values along with an optional histogram of the values as specified by the bucket options.

LogSink
Describes a sink used to export log entries to one of the following destinations in any project: a Cloud Storage bucket, a BigQuery dataset, or a Cloud Pub/Sub topic. A logs filter controls which log entries are exported. The sink must be created within a project, organization, billing account, or folder.

Logging
Central instance to access all Logging related resource activities

MethodInfo
Contains information about an API request.

MetricDescriptor
Defines a metric type and its schema. Once a metric descriptor is created, deleting or altering it stops data collection and makes the metric type's existing data unusable.

MetricDescriptorMetadata
Additional annotations that can be used to guide the usage of a metric.

MonitoredResource
An object representing a resource that can be used for monitoring, logging, billing, or other purposes. Examples include virtual machine instances, databases, and storage devices such as disks. The type field identifies a MonitoredResourceDescriptor object that describes the resource's schema. Information in the labels field identifies the actual resource and its attributes according to the schema. For example, a particular Compute Engine VM instance could be represented by the following object, because the MonitoredResourceDescriptor for "gce_instance" has labels "instance_id" and "zone": { "type": "gce_instance", "labels": { "instance_id": "12345678901234", "zone": "us-central1-a" }}

MonitoredResourceDescriptor
An object that describes the schema of a MonitoredResource object using a type name and a set of labels. For example, the monitored resource descriptor for Google Compute Engine VM instances has a type of "gce_instance" and specifies the use of the labels "instance_id" and "zone" to identify particular VM instances.Different APIs can support different monitored resource types. APIs generally provide a list method that returns the monitored resource descriptors used by the API.

MonitoredResourceDescriptorListCall
Lists the descriptors for monitored resource types used by Logging.

MonitoredResourceDescriptorMethods
A builder providing access to all methods supported on monitoredResourceDescriptor resources. It is not used directly, but through the Logging hub.

MonitoredResourceMetadata
Auxiliary metadata for a MonitoredResource object. MonitoredResource objects contain the minimum set of information to uniquely identify a monitored resource instance. There is some other useful auxiliary metadata. Monitoring and Logging use an ingestion pipeline to extract metadata for cloud resources of all types, and store the metadata in this message.

MultiPartReader
Provides a Read interface that converts multiple parts into the protocol identified by RFC2387. Note: This implementation is just as rich as it needs to be to perform uploads to google APIs, and might not be a fully-featured implementation.

OrganizationExclusionCreateCall
Creates a new exclusion in a specified parent resource. Only log entries belonging to that resource can be excluded. You can have up to 10 exclusions in a resource.

OrganizationExclusionDeleteCall
Deletes an exclusion.

OrganizationExclusionGetCall
Gets the description of an exclusion.

OrganizationExclusionListCall
Lists all the exclusions in a parent resource.

OrganizationExclusionPatchCall
Changes one or more properties of an existing exclusion.

OrganizationLogDeleteCall
Deletes all the log entries in a log. The log reappears if it receives new entries. Log entries written shortly before the delete operation might not be deleted.

OrganizationLogListCall
Lists the logs in projects, organizations, folders, or billing accounts. Only logs that have entries are listed.

OrganizationMethods
A builder providing access to all methods supported on organization resources. It is not used directly, but through the Logging hub.

OrganizationSinkCreateCall
Creates a sink that exports specified log entries to a destination. The export of newly-ingested log entries begins immediately, unless the sink's writer_identity is not permitted to write to the destination. A sink can export log entries only from the resource owning the sink.

OrganizationSinkDeleteCall
Deletes a sink. If the sink has a unique writer_identity, then that service account is also deleted.

OrganizationSinkGetCall
Gets a sink.

OrganizationSinkListCall
Lists sinks.

OrganizationSinkPatchCall
Updates a sink. This method replaces the following fields in the existing sink with values from the new sink: destination, and filter.The updated sink might also have a new writer_identity; see the unique_writer_identity field.

OrganizationSinkUpdateCall
Updates a sink. This method replaces the following fields in the existing sink with values from the new sink: destination, and filter.The updated sink might also have a new writer_identity; see the unique_writer_identity field.

ProjectExclusionCreateCall
Creates a new exclusion in a specified parent resource. Only log entries belonging to that resource can be excluded. You can have up to 10 exclusions in a resource.

ProjectExclusionDeleteCall
Deletes an exclusion.

ProjectExclusionGetCall
Gets the description of an exclusion.

ProjectExclusionListCall
Lists all the exclusions in a parent resource.

ProjectExclusionPatchCall
Changes one or more properties of an existing exclusion.

ProjectLogDeleteCall
Deletes all the log entries in a log. The log reappears if it receives new entries. Log entries written shortly before the delete operation might not be deleted.

ProjectLogListCall
Lists the logs in projects, organizations, folders, or billing accounts. Only logs that have entries are listed.

ProjectMethods
A builder providing access to all methods supported on project resources. It is not used directly, but through the Logging hub.

ProjectMetricCreateCall
Creates a logs-based metric.

ProjectMetricDeleteCall
Deletes a logs-based metric.

ProjectMetricGetCall
Gets a logs-based metric.

ProjectMetricListCall
Lists logs-based metrics.

ProjectMetricUpdateCall
Creates or updates a logs-based metric.

ProjectSinkCreateCall
Creates a sink that exports specified log entries to a destination. The export of newly-ingested log entries begins immediately, unless the sink's writer_identity is not permitted to write to the destination. A sink can export log entries only from the resource owning the sink.

ProjectSinkDeleteCall
Deletes a sink. If the sink has a unique writer_identity, then that service account is also deleted.

ProjectSinkGetCall
Gets a sink.

ProjectSinkListCall
Lists sinks.

ProjectSinkPatchCall
Updates a sink. This method replaces the following fields in the existing sink with values from the new sink: destination, and filter.The updated sink might also have a new writer_identity; see the unique_writer_identity field.

ProjectSinkUpdateCall
Updates a sink. This method replaces the following fields in the existing sink with values from the new sink: destination, and filter.The updated sink might also have a new writer_identity; see the unique_writer_identity field.

RangeResponseHeader

ResumableUploadHelper
A utility type to perform a resumable upload from start to end.

ServerError

ServerMessage

SinkCreateCall
Creates a sink that exports specified log entries to a destination. The export of newly-ingested log entries begins immediately, unless the sink's writer_identity is not permitted to write to the destination. A sink can export log entries only from the resource owning the sink.

SinkDeleteCall
Deletes a sink. If the sink has a unique writer_identity, then that service account is also deleted.

SinkGetCall
Gets a sink.

SinkListCall
Lists sinks.

SinkMethods
A builder providing access to all methods supported on sink resources. It is not used directly, but through the Logging hub.

SinkUpdateCall
Updates a sink. This method replaces the following fields in the existing sink with values from the new sink: destination, and filter.The updated sink might also have a new writer_identity; see the unique_writer_identity field.

WriteLogEntriesRequest
The parameters to WriteLogEntries.

WriteLogEntriesResponse
Result returned from WriteLogEntries. empty

XUploadContentType
The X-Upload-Content-Type header.

Enums

Error

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.

Traits

CallBuilder
Identifies types which represent builders for a particular resource method

Delegate
A trait specifying functionality to help controlling any request performed by the API. The trait has a conservative default implementation.

Hub
Identifies the Hub. There is only one per library, this trait is supposed to make intended use more explicit. The hub allows to access all resource methods more easily.

MethodsBuilder
Identifies types for building methods of a particular resource type

NestedType
Identifies types which are only used by other types internally. They have no special meaning, this trait just marks them for completeness.

Part
Identifies types which are only used as part of other types, which usually are carrying the Resource trait.

ReadSeek
A utility to specify reader types which provide seeking capabilities too

RequestValue
Identifies types which are used in API requests.

Resource
Identifies types which can be inserted and deleted. Types with this trait are most commonly used by clients of this API.

ResponseResult
Identifies types which are used in API responses.

ToParts
A trait for all types that can convert themselves into a parts string

UnusedType
Identifies types which are not actually used by the API This might be a bug within the google API schema.

Functions

remove_json_null_values

Type Definitions

Result
A universal result type used as return for all calls.